+++ src/renderer.js
@@ -4,3 +4,4 @@ export { default as createRepresentativeRenderer } from './representative-render
 export { default as createSvgRenderer } from './svg-renderer';
 export { default as createD3Renderer } from './d3-renderer';
 export { default as createVegaLiteRenderer } from './vega-lite-renderer';
+export { default as createObservablePlotRenderer } from './observable-plot-renderer';